requirements.txt
Django==5.1.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.11
python-dotenv==1.0.1
pytest==8.0.2
pytest-django==4.8.0
drf-yasg==1.21.7
setuptools<81
